% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosnapshot.R
\name{autosnapshot}
\alias{autosnapshot}
\title{Automatically stage all changes and commit}
\usage{
autosnapshot(repopath = getwd(), autobranch = NULL)
}
\arguments{
\item{repopath}{Character. Specify path to existing repository.}

\item{autobranch}{Character. Specify the name of an existing branch. Only if HEAD is pointing to this branch an auto-snapshot will be created. Exception: Choose NULL (default) to ignore branch-dependency.}
}
\description{
This function will stage all changes and commit a snapshot with a predefined message (including user and timestamp) via git2r.
}
\seealso{
\code{\link[git2r]{repository}}, \code{\link[git2r]{status}}, \code{\link[git2r]{add}}, \code{\link[git2r]{commit}}
}
