% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{Copy a directory into timestamp-dir}
\usage{
backup(target_dir_backup = "project_subdir",
  source_dir_backup = file.path(getwd()), rebuild_target_dir_backup = FALSE,
  exclude_dir_backup = FALSE, exclude_files_backup = FALSE)
}
\arguments{
\item{target_dir_backup}{Character. Path to the target directory. Alternatively use
'project_subdir' to point to a subfolder of your current working directory.}

\item{source_dir_backup}{Character. Path to the source directory. Default is the
current working directory.}

\item{rebuild_target_dir_backup}{Logical. If TRUE all directories within the target
directory will be deleted! Only the current version of your backup will be
existent.}

\item{exclude_dir_backup}{Character. Specify directories to exclude from
backup. If you want to exclude multiple directories use | as separator.
Choose FALSE if you don not want to exclude any directories. E.g. 'packrat|.git|in/data|out|.cache'}

\item{exclude_files_backup}{Character. Specify files to exclude from backup. If you
want to exclude multiple directories use | as separator. Choose FALSE if
you don not want to exclude any files. E.g. '*.RData|*.Rhistory'}
}
\description{
The function will copy the specified directory and create a
  directory at the target. The target directory will be named after the
  current date and time.
}
\note{
Beware! If you choose an existing target directory and option
  rebuild_target_dir_backup = TRUE: ALL EXISTING FILES WILL BE DELETED!
}
\author{
Frederik Sachser
}
