% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_write.R
\name{cache}
\alias{cache}
\title{Cache a value}
\usage{
cache(name, value)
}
\arguments{
\item{name}{The cache name}

\item{value}{The value to cache (will be evaluated if it's an expression)}
}
\value{
The cached value
}
\description{
Cache a value
}
