% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{capture_output}
\alias{capture_output}
\title{Capture console output and errors from an expression}
\usage{
capture_output(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
List containing result (NULL if successful, error if failed) and output
}
\description{
Capture console output and errors from an expression
}
