% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_instructions.R
\name{check_instructions}
\alias{check_instructions}
\title{check_instructions}
\usage{
check_instructions(cache_dir = ".cache")
}
\description{
check_instructions
}
\note{
This function is part of a family of functions each of which end with 
  '_instructions'. The order to call these functions is: 
  'prepare_instructions', 'implement_instructions', 'check_instructions', 
  'execute_instructions' and optionally 'output_instructions'. There is a
  wrapper for these functions called 'instructions'.
}
\author{
Frederik Sachser
}

