% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cli.R
\name{cli_uninstall}
\alias{cli_uninstall}
\alias{uninstall_cli}
\title{Uninstall Framework CLI Tool}
\usage{
cli_uninstall(location = c("user", "system"))

uninstall_cli(location = c("user", "system"))
}
\arguments{
\item{location}{Installation location to remove: "user" or "system".
Must match the location used during installation.}
}
\description{
Removes the global \code{framework} command.
}
\examples{
\dontrun{
# Uninstall user installation
cli_uninstall()

# Uninstall system installation
cli_uninstall(location = "system")
}

}
