% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_settings}
\alias{cloud_settings}
\title{Fetch Your Cloud Settings}
\usage{
cloud_settings()
}
\value{
A list with \code{revision}, \code{schema_version}, and \code{document}.
}
\description{
Retrieves your settings document from framework.pub. These settings are
overlaid on local defaults when creating projects with \code{\link[=new]{new()}}.
}
