% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connection_get}
\alias{connection_get}
\title{Get a database connection from config}
\usage{
connection_get(name)
}
\arguments{
\item{name}{Character. Name of the connection in config.yml (e.g., "postgres")}
}
\value{
A database connection object (DBIConnection)
}
\description{
Gets a database connection based on the connection name in config.yml.
}
\examples{
\dontrun{
conn <- connection_get("postgres")
DBI::dbListTables(conn)
}

}
