% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{connection_restore}
\alias{connection_restore}
\title{Restore a soft-deleted record}
\usage{
connection_restore(conn, table_name, id)
}
\arguments{
\item{conn}{Database connection}

\item{table_name}{Name of the table}

\item{id}{The ID of the record to restore}
}
\value{
Number of rows affected
}
\description{
Restores a soft-deleted record by setting deleted_at to NULL.
Only works on tables with a deleted_at column.
}
\examples{
\dontrun{
conn <- connection_get("postgres")
connection_restore(conn, "users", 1)
DBI::dbDisconnect(conn)
}

}
