% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_inst.R
\name{copy_inst}
\alias{copy_inst}
\title{Copy installed files}
\usage{
copy_inst(pkg = "framework", pkgdir = "templates", targetdir = "R")
}
\arguments{
\item{pkg}{Character. Package name.}

\item{pkgdir}{Character. Sub directory of "inst".}

\item{targetdir}{Character. Path to target directory}
}
\description{
Copy installed files
}
