% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{db_execute}
\alias{db_execute}
\title{Execute a database statement}
\usage{
db_execute(query, connection_name, ...)
}
\arguments{
\item{query}{SQL statement to execute}

\item{connection_name}{Name of the connection in config.yml}

\item{...}{Additional arguments passed to DBI::dbExecute}
}
\value{
Number of rows affected
}
\description{
Executes a SQL statement on a database without returning results. The connection
is created, used, and automatically closed.
}
\examples{
\donttest{
if (FALSE) {
rows <- db_execute("DELETE FROM cache WHERE expired = TRUE", "my_db")
}
}

}
