% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_deprecated_instructions.R
\name{delete_deprecated_instructions}
\alias{delete_deprecated_instructions}
\title{delete_deprecated_instructions}
\usage{
delete_deprecated_instructions(cache_dir = ".cache")
}
\arguments{
\item{cache_dir}{Character. Specify file path for the cache directory.}
}
\description{
delete_deprecated_instructions
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{prepare_instructions}}, 
  \code{\link{implement_instructions}}, \code{\link{check_instructions}}, 
  \code{\link{instructions}}, \code{\link{execute_instructions}},
  \code{\link{output_instructions}}
}

