% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_skills.R
\name{.ai_skills_for_type}
\alias{.ai_skills_for_type}
\title{Skills that apply to a given project type}
\usage{
.ai_skills_for_type(project_type)
}
\description{
Skills that apply to a given project type
}
\keyword{internal}
