% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.archive_init_file}
\alias{.archive_init_file}
\title{Archive init.R after successful initialization}
\usage{
.archive_init_file(subdir = NULL)
}
\description{
Archive init.R after successful initialization
}
\keyword{internal}
