% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.build_directory_table}
\alias{.build_directory_table}
\title{Build directory table from config}
\usage{
.build_directory_table(dirs, project_type)
}
\description{
Build directory table from config
}
\keyword{internal}
