% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{.calculate_file_hash}
\alias{.calculate_file_hash}
\title{Calculate hash of a file}
\usage{
.calculate_file_hash(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
The hash of the file as a character string
}
\description{
Calculate hash of a file
}
\keyword{internal}
