% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_catalog.R
\name{.catalog_find_field}
\alias{.catalog_find_field}
\title{Convenience accessor for catalog field definitions}
\usage{
.catalog_find_field(catalog, field_id)
}
\description{
Convenience accessor for catalog field definitions
}
\keyword{internal}
