% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_audit.R
\name{.check_git_history}
\alias{.check_git_history}
\title{Check git history for leaked data files}
\usage{
.check_git_history(data_dirs, extensions, history_depth, verbose)
}
\description{
Check git history for leaked data files
}
\keyword{internal}
