% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_mysql.R
\name{.check_mysql_exists}
\alias{.check_mysql_exists}
\title{Check if a MySQL/MariaDB database exists}
\usage{
.check_mysql_exists(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if database exists, FALSE otherwise
}
\description{
Check if a MySQL/MariaDB database exists
}
\keyword{internal}
