% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_postgres.R
\name{.check_postgres_exists}
\alias{.check_postgres_exists}
\title{Check if a PostgreSQL database exists}
\usage{
.check_postgres_exists(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if database exists, FALSE otherwise
}
\description{
Check if a PostgreSQL database exists
}
\keyword{internal}
