% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlite.R
\name{.check_sqlite_exists}
\alias{.check_sqlite_exists}
\title{Check if a SQLite database exists}
\usage{
.check_sqlite_exists(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if database exists, FALSE otherwise
}
\description{
Check if a SQLite database exists
}
\keyword{internal}
