% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlserver.R
\name{.check_sqlserver_exists}
\alias{.check_sqlserver_exists}
\title{Check if a SQL Server database exists}
\usage{
.check_sqlserver_exists(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if database exists, FALSE otherwise
}
\description{
Check if a SQL Server database exists
}
\keyword{internal}
