% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_duckdb.R
\name{.connect_duckdb}
\alias{.connect_duckdb}
\title{Connect to a DuckDB database}
\usage{
.connect_duckdb(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
A DuckDB database connection
}
\description{
Connect to a DuckDB database
}
\keyword{internal}
