% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlite.R
\name{.connect_sqlite}
\alias{.connect_sqlite}
\title{Connect to a SQLite database}
\usage{
.connect_sqlite(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
A SQLite database connection
}
\description{
Connect to a SQLite database
}
\keyword{internal}
