% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlserver.R
\name{.connect_sqlserver}
\alias{.connect_sqlserver}
\title{Connect to a SQL Server database}
\usage{
.connect_sqlserver(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
A SQL Server database connection via ODBC
}
\description{
Connect to a SQL Server database
}
\keyword{internal}
