% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{.copy_config_files}
\alias{.copy_config_files}
\title{Copy config files between directories}
\usage{
.copy_config_files(from_dir, to_dir, overwrite = FALSE)
}
\arguments{
\item{from_dir}{Source directory}

\item{to_dir}{Destination directory}

\item{overwrite}{Whether to overwrite existing files (default FALSE)}
}
\description{
Copy config files between directories
}
\keyword{internal}
