% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.create_code_workspace}
\alias{.create_code_workspace}
\title{Create .code-workspace file for VSCode/Positron}
\usage{
.create_code_workspace(project_dir, name)
}
\description{
Create .code-workspace file for VSCode/Positron
}
\keyword{internal}
