% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.create_config_file}
\alias{.create_config_file}
\title{Create settings.yml from template}
\usage{
.create_config_file(type = "analysis", attach_defaults = TRUE, subdir = NULL)
}
\description{
Create settings.yml from template
}
\keyword{internal}
