% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_duckdb.R
\name{.create_duckdb_db}
\alias{.create_duckdb_db}
\title{Create a new DuckDB database}
\usage{
.create_duckdb_db(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if successful
}
\description{
Create a new DuckDB database
}
\keyword{internal}
