% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ide_config.R
\name{.create_ide_configs}
\alias{.create_ide_configs}
\title{Create IDE configuration files}
\usage{
.create_ide_configs(project_name, target_dir = ".", python = FALSE)
}
\arguments{
\item{project_name}{Project name}

\item{target_dir}{Target directory}

\item{python}{Include Python configuration}
}
\description{
Create IDE configuration files
}
\keyword{internal}
