% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.create_init_file}
\alias{.create_init_file}
\title{Create init.R from template}
\usage{
.create_init_file(project_name, type, lintr, styler, subdir = NULL)
}
\description{
Create init.R from template
}
\keyword{internal}
