% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_mysql.R
\name{.create_mysql_db}
\alias{.create_mysql_db}
\title{Create a new MySQL/MariaDB database}
\usage{
.create_mysql_db(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if successful
}
\description{
Create a new MySQL/MariaDB database
}
\keyword{internal}
