% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_postgres.R
\name{.create_postgres_db}
\alias{.create_postgres_db}
\title{Create a new PostgreSQL database}
\usage{
.create_postgres_db(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if successful
}
\description{
Create a new PostgreSQL database
}
\keyword{internal}
