% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.create_project_config}
\alias{.create_project_config}
\title{Create project config.yml}
\usage{
.create_project_config(
  project_dir,
  name,
  type,
  author,
  packages,
  directories,
  extra_directories,
  ai,
  git,
  scaffold,
  settings_dir,
  connections = NULL,
  connections_file = NULL,
  render_dirs = NULL,
  quarto = NULL
)
}
\description{
Create project config.yml
}
\keyword{internal}
