% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.create_project_directories}
\alias{.create_project_directories}
\title{Create project subdirectories}
\usage{
.create_project_directories(
  project_dir,
  directories,
  extra_directories,
  render_dirs = NULL
)
}
\description{
Create project subdirectories
}
\keyword{internal}
