% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlite.R
\name{.create_sqlite_db}
\alias{.create_sqlite_db}
\title{Create a new SQLite database}
\usage{
.create_sqlite_db(config)
}
\arguments{
\item{config}{Connection configuration from settings.yml}
}
\value{
TRUE if successful
}
\description{
Create a new SQLite database
}
\keyword{internal}
