% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_sqlserver.R
\name{.create_sqlserver_db}
\alias{.create_sqlserver_db}
\title{Create a new SQL Server database}
\usage{
.create_sqlserver_db(config)
}
\arguments{
\item{config}{Connection configuration from config.yml}
}
\value{
TRUE if successful
}
\description{
Create a new SQL Server database
}
\keyword{internal}
