% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.create_stub_files}
\alias{.create_stub_files}
\title{Create stub files for specific project types}
\usage{
.create_stub_files(project_dir, type, name, author)
}
\description{
Create stub files for specific project types
}
\keyword{internal}
