% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ide_config.R
\name{.create_vscode_settings}
\alias{.create_vscode_settings}
\title{Create VS Code settings.json}
\usage{
.create_vscode_settings(target_dir = ".", python = FALSE)
}
\arguments{
\item{target_dir}{Target directory (defaults to current)}

\item{python}{Include Python configuration}
}
\description{
Create VS Code settings.json
}
\keyword{internal}
