% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_encrypt.R
\name{.decrypt_data}
\alias{.decrypt_data}
\title{Decrypt data using sodium}
\usage{
.decrypt_data(encrypted_data, key)
}
\arguments{
\item{encrypted_data}{Encrypted data as raw}

\item{key}{Encryption key}
}
\value{
Decrypted data
}
\description{
Decrypt data using sodium
}
\keyword{internal}
