% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.default_root_render_dir_for_type}
\alias{.default_root_render_dir_for_type}
\title{Resolve default root render_dir (if defined) for a project type}
\usage{
.default_root_render_dir_for_type(type)
}
\description{
Resolve default root render_dir (if defined) for a project type
}
\keyword{internal}
