% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.df_to_list_of_lists}
\alias{.df_to_list_of_lists}
\title{Convert data frame to list of lists for YAML serialization
JSON arrays of objects become data frames in R, but YAML needs list of lists}
\usage{
.df_to_list_of_lists(df)
}
\description{
Convert data frame to list of lists for YAML serialization
JSON arrays of objects become data frames in R, but YAML needs list of lists
}
\keyword{internal}
