% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.display_next_steps}
\alias{.display_next_steps}
\title{Display next steps after initialization}
\usage{
.display_next_steps(project_name = NULL, type = "project", use_renv = FALSE)
}
\description{
Display next steps after initialization
}
\keyword{internal}
