% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_encrypt.R
\name{.encrypt_data}
\alias{.encrypt_data}
\title{Encrypt data using sodium}
\usage{
.encrypt_data(value, key)
}
\arguments{
\item{value}{Data to encrypt}

\item{key}{Encryption key}
}
\value{
Encrypted data as raw
}
\description{
Encrypt data using sodium
}
\keyword{internal}
