% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_catalog.R
\name{.framework_catalog_user_path}
\alias{.framework_catalog_user_path}
\title{Get the path to the user-editable settings catalog override}
\usage{
.framework_catalog_user_path()
}
\description{
Get the path to the user-editable settings catalog override
}
\keyword{internal}
