% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.generate_data_section}
\alias{.generate_data_section}
\title{Generate Data Management section}
\usage{
.generate_data_section(config, project_type)
}
\description{
Generate Data Management section
}
\keyword{internal}
