% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.generate_environment_section}
\alias{.generate_environment_section}
\title{Generate Framework Environment section}
\usage{
.generate_environment_section(config, project_type)
}
\description{
Generate Framework Environment section
}
\keyword{internal}
