% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.generate_function_reference}
\alias{.generate_function_reference}
\title{Generate Function Reference section}
\usage{
.generate_function_reference()
}
\description{
Generate Function Reference section
}
\keyword{internal}
