% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{.generate_hook_script}
\alias{.generate_hook_script}
\title{Generate pre-commit hook script}
\usage{
.generate_hook_script(ai_sync_enabled, data_security_enabled)
}
\description{
Generate pre-commit hook script
}
\keyword{internal}
