% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.generate_packages_section}
\alias{.generate_packages_section}
\title{Generate Installed Packages section}
\usage{
.generate_packages_section(config)
}
\description{
Generate Installed Packages section
}
\keyword{internal}
