% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_read.R
\name{.get_cache}
\alias{.get_cache}
\title{Get a cache value}
\usage{
.get_cache(name)
}
\arguments{
\item{name}{The cache name}
}
\value{
The cached result, or NULL if not found or hash mismatch
}
\description{
Get a cache value
}
\keyword{internal}
