% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_audit.R
\name{.get_data_directories}
\alias{.get_data_directories}
\title{Get data directories from config}
\usage{
.get_data_directories(config, verbose)
}
\description{
Get data directories from config
}
\keyword{internal}
