% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{.get_data_record}
\alias{.get_data_record}
\title{Get a data value}
\usage{
.get_data_record(name)
}
\arguments{
\item{name}{The data name}
}
\value{
The data metadata (hash), or NULL if not found
}
\description{
Get a data value
}
\keyword{internal}
