% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.get_example_data_path}
\alias{.get_example_data_path}
\title{Get example data path based on config}
\usage{
.get_example_data_path(dirs, project_type, path_type)
}
\description{
Get example data path based on config
}
\keyword{internal}
