% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.get_package_requirements}
\alias{.get_package_requirements}
\title{Get package requirements from config}
\usage{
.get_package_requirements(config)
}
\arguments{
\item{config}{Configuration object from config::get()}
}
\description{
Get package requirements from config
}
\keyword{internal}
