% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{.get_placeholders}
\alias{.get_placeholders}
\title{Generate database-appropriate parameter placeholders}
\usage{
.get_placeholders(conn, n)
}
\arguments{
\item{conn}{Database connection}

\item{n}{Number of placeholders needed}
}
\value{
Character vector of placeholders
}
\description{
Generate database-appropriate parameter placeholders
}
\keyword{internal}
