% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{.git_available}
\alias{.git_available}
\title{Check if git is available on the system}
\usage{
.git_available()
}
\value{
TRUE if git is available, FALSE otherwise
}
\description{
Check if git is available on the system
}
\keyword{internal}
