% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_s3.R
\name{.guess_content_type}
\alias{.guess_content_type}
\title{Guess content type from file extension}
\usage{
.guess_content_type(file)
}
\arguments{
\item{file}{Character. File path.}
}
\value{
Character. MIME type.
}
\description{
Guess content type from file extension
}
\keyword{internal}
