% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.install_package}
\alias{.install_package}
\title{Install a package if not already installed}
\usage{
.install_package(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\description{
Install a package if not already installed
}
\keyword{internal}
