% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.install_required_packages}
\alias{.install_required_packages}
\title{Install required packages from config}
\usage{
.install_required_packages(config)
}
\arguments{
\item{config}{Configuration object from read_config()}
}
\description{
Install required packages from config
}
\keyword{internal}
