% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_notebook.R
\name{.list_available_stubs}
\alias{.list_available_stubs}
\title{List Available Stub Templates}
\usage{
.list_available_stubs(ext = NULL)
}
\arguments{
\item{ext}{Character. File extension to filter by}
}
\value{
Character vector of stub names
}
\description{
List Available Stub Templates
}
\keyword{internal}
