% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.load_configuration}
\alias{.load_configuration}
\title{Load configuration from config.yml}
\usage{
.load_configuration(config_file = "config.yml")
}
\description{
Load configuration from config.yml
}
\keyword{internal}
