% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.load_functions}
\alias{.load_functions}
\title{Load all R files from functions directories}
\usage{
.load_functions(config_file = NULL)
}
\description{
Load all R files from functions directories
}
\keyword{internal}
