% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.load_libraries}
\alias{.load_libraries}
\title{Load all libraries specified in config}
\usage{
.load_libraries(config)
}
\arguments{
\item{config}{Configuration object from read_config()}
}
\description{
Load all libraries specified in config
}
\keyword{internal}
