% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{.load_template_content}
\alias{.load_template_content}
\title{Load template content from inst/templates}
\usage{
.load_template_content(template_name)
}
\description{
Load template content from inst/templates
}
\keyword{internal}
