% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption_core.R
\docType{data}
\name{.MAGIC_BYTES}
\alias{.MAGIC_BYTES}
\title{Encryption core functions for Framework}
\format{
An object of class \code{raw} of length 6.
}
\usage{
.MAGIC_BYTES
}
\description{
Password-based encryption using libsodium (Ansible Vault style).
Derives cryptographic keys from user passwords using PBKDF2.
}
\keyword{internal}
