% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_notebook.R
\name{.normalize_notebook_name}
\alias{.normalize_notebook_name}
\title{Normalize Notebook Name and Detect Type}
\usage{
.normalize_notebook_name(name, type = c("quarto", "rmarkdown", "script"))
}
\arguments{
\item{name}{Character. File name with or without extension}

\item{type}{Character. Type preference}
}
\value{
List with name, type, and ext
}
\description{
Normalize Notebook Name and Detect Type
}
\keyword{internal}
