% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_context.R
\name{.parse_sections}
\alias{.parse_sections}
\title{Parse markdown into sections based on ## headings}
\usage{
.parse_sections(content)
}
\description{
Parse markdown into sections based on ## headings
}
\keyword{internal}
