% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.prompt_project_config}
\alias{.prompt_project_config}
\title{Prompt user for project configuration}
\usage{
.prompt_project_config(project_name, project_structure, lintr, styler)
}
\arguments{
\item{project_name}{Project name (NULL to prompt)}

\item{project_structure}{Project structure (NULL to prompt)}

\item{lintr}{Lintr style (NULL to prompt)}

\item{styler}{Styler style (NULL to prompt)}
}
\value{
List of configuration parameters
}
\description{
Prompt user for project configuration
}
\keyword{internal}
