% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_delete.R
\name{.remove_cache}
\alias{.remove_cache}
\title{Remove a cache value}
\usage{
.remove_cache(name)
}
\arguments{
\item{name}{The cache name to remove}
}
\description{
Remove a cache value
}
\keyword{internal}
