% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_db.R
\name{.remove_metadata}
\alias{.remove_metadata}
\title{Remove a metadata value}
\usage{
.remove_metadata(key)
}
\arguments{
\item{key}{The metadata key to remove}
}
\description{
Remove a metadata value
}
\keyword{internal}
