% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_s3.R
\name{.s3_public_url}
\alias{.s3_public_url}
\title{Generate public URL for S3 object}
\usage{
.s3_public_url(key, s3_config)
}
\arguments{
\item{key}{Character. Object key in S3.}

\item{s3_config}{List. S3 configuration.}
}
\value{
Character. Public URL.
}
\description{
Generate public URL for S3 object
}
\keyword{internal}
