% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_write.R
\name{.set_cache}
\alias{.set_cache}
\title{Set a cache value}
\usage{
.set_cache(name, value)
}
\arguments{
\item{name}{The cache name}

\item{value}{The value to cache (will be evaluated if it's an expression)}
}
\description{
Set a cache value
}
\keyword{internal}
