% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_write.R
\name{.set_data}
\alias{.set_data}
\title{Set a data value}
\usage{
.set_data(name, encrypted = FALSE, key = NULL, hash = NULL)
}
\arguments{
\item{name}{The data name}

\item{encrypted}{Whether the data is encrypted}

\item{key}{The encryption key}

\item{hash}{The hash of the data}
}
\description{
Set a data value
}
\keyword{internal}
