% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_db.R
\name{.set_metadata}
\alias{.set_metadata}
\title{Set a metadata value}
\usage{
.set_metadata(key, value)
}
\arguments{
\item{key}{The metadata key}

\item{value}{The metadata value}
}
\description{
Set a metadata value
}
\keyword{internal}
