% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_migrate.R
\name{.settings_need_migration}
\alias{.settings_need_migration}
\title{Check if settings need migration}
\usage{
.settings_need_migration(settings)
}
\arguments{
\item{settings}{List containing settings}
}
\value{
Logical indicating whether migration is needed
}
\description{
Check if settings need migration
}
\keyword{internal}
