% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_write.R
\name{.update_data}
\alias{.update_data}
\title{Update data hash and timestamp}
\usage{
.update_data(name, hash, type)
}
\arguments{
\item{name}{The data name}

\item{hash}{The new hash}

\item{type}{The data type}
}
\description{
Update data hash and timestamp
}
\keyword{internal}
