% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.update_gitignore_for_renv}
\alias{.update_gitignore_for_renv}
\title{Update .gitignore for renv}
\usage{
.update_gitignore_for_renv()
}
\value{
Invisibly returns NULL
}
\description{
Adds renv-related entries to .gitignore if they don't already exist.
}
\keyword{internal}
