% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{.update_hook_config}
\alias{.update_hook_config}
\title{Update hook configuration in config.yml}
\usage{
.update_hook_config(hook_name, enabled, config_file)
}
\description{
Update hook configuration in config.yml
}
\keyword{internal}
