% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_read.R
\name{.validate_refresh}
\alias{.validate_refresh}
\title{Validate refresh parameter}
\usage{
.validate_refresh(refresh)
}
\arguments{
\item{refresh}{Boolean or function that returns boolean}
}
\value{
Boolean indicating if refresh is needed
}
\description{
Validate refresh parameter
}
\keyword{internal}
