% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_instructions.R
\name{execute_instructions}
\alias{execute_instructions}
\title{Execute instructions of df_cache.rds}
\usage{
execute_instructions(filename_df_cache = ".cache/df_cache.rds")
}
\arguments{
\item{filename_df_cache}{Character. Specify the file path where to find the 
file df_cache.rds.}
}
\description{
The function executes the instructions according to the file 
  df_cache.rds. Furthermore it is a wrapper for specify_instructions.
}
\note{
The function only makes sense after calling 'instructions()'. You might
  want to create a new framework-project (using 'project_framework()') and 
  have a look at the file make.R.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{specify_instructions}}
}

