% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{framework version of file.edit}
\usage{
fe(file, ...)
}
\arguments{
\item{file}{Character or integer. Specify full file path to open or create a file. If the value is an integer and 'framework_params' is in your search path it will look up the index of the input_file (at the point in time of the last processing) to open it.}

\item{...}{Arguments passed to 'utile::file_edit'.}
}
\description{
This function will open existing files or create a new one by calling 'framework::template_rmd'.
}
\seealso{
\code{\link[utils]{file.edit}}, \code{\link{template_rmd}}
}
