% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-package.R
\docType{package}
\name{framework-package}
\alias{framework}
\alias{framework-package}
\title{framework: Structured Data Science Project Scaffolding}
\description{
Project scaffolding and workflow tools for reproducible data science. Manages packages, tracks data integrity, handles database connections, generates notebooks, and publishes to S3-compatible storage. More information at \url{https://framework.table1.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://framework.table1.org}
  \item \url{https://github.com/table1/framework}
  \item Report bugs at \url{https://github.com/table1/framework/issues}
}

}
\author{
\strong{Maintainer}: Erik Westlund \email{erik@table1.org} [copyright holder]

Authors:
\itemize{
  \item Erik Westlund \email{erik@table1.org} [copyright holder]
}

}
\keyword{internal}
