% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_read.R
\name{get_and_cache}
\alias{get_and_cache}
\title{Get a value, caching the result}
\usage{
get_and_cache(name, expr)
}
\arguments{
\item{name}{The cache name}

\item{expr}{The expression to evaluate and cache}
}
\value{
The result of expr
}
\description{
Get a value, caching the result
}
