% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{get_data_spec}
\alias{get_data_spec}
\title{Get data specification from config}
\usage{
get_data_spec(path)
}
\arguments{
\item{path}{Dot notation path (e.g. "source.private.example")}
}
\value{
The data specification as a list
}
\description{
Gets the data specification for a given dot notation path
}
