% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{get_default_global_config}
\alias{get_default_global_config}
\title{Get default global configuration structure}
\usage{
get_default_global_config()
}
\value{
List with default global configuration
}
\description{
Get default global configuration structure
}
\keyword{internal}
