% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_query}
\alias{get_query}
\title{Get data from a database query}
\usage{
get_query(query, connection_name, ...)
}
\arguments{
\item{query}{SQL query to execute}

\item{connection_name}{Name of the connection in config.yml}

\item{...}{Additional arguments passed to DBI::dbGetQuery}
}
\value{
A data frame with the query results
}
\description{
Gets data from a database using a query and connection name.
}
