% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_result}
\alias{get_result}
\title{Get a result}
\usage{
get_result(name)
}
\arguments{
\item{name}{Result name}
}
\value{
The result value, or NULL if not found or hash mismatch
}
\description{
Get a result
}
