% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_init.R
\name{git_init}
\alias{git_init}
\title{Initialize a git repo}
\usage{
git_init()
}
\value{
Add and commit all existing files to branch master. Afterwards checkout to a newly created branch devel.
}
\description{
Initialize a git repo within the current working directory.
}
\note{
The function will work only if git is installed and '.git' is absent.
}
\author{
Frederik Sachser
}

