% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui_restart}
\alias{gui_restart}
\title{Restart the GUI server}
\usage{
gui_restart(port = 8080, browse = FALSE)
}
\arguments{
\item{port}{Port number to use}

\item{browse}{Automatically open browser}
}
\description{
Restart the GUI server
}
