% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_instructions.R
\name{implement_instructions}
\alias{implement_instructions}
\title{Implement specified instructions}
\usage{
implement_instructions(source_files, spin_index, cache_index, cache_dir,
  source_dir, target_dir_figure, target_dir_docs)
}
\arguments{
\item{source_files}{Character. Specify a vector of file paths, if you want to
create predefined templates for your analysis (relative to top-level). 
Use file extensions '.R' or '.Rmd'.}

\item{spin_index}{Integer. Can be of length one or a vector. Specify index of
'source_files' that should  be spinned (ignored for file extension 'Rmd'). 
Choose 0 (zero) if you do not want to spin any R-files. Choose 999 if you 
want to spin all R-files.}

\item{cache_index}{Integer. Can be of length one or a vector. Specify index 
of 'source_files' that should  be integrated into the cache. Choose 0 
(zero) if you do not want to use the cache. Choose 999 if you want to use 
the cache for all files.}

\item{cache_dir}{Character. Specify file path for the cache directory.}

\item{source_dir}{Character. Specify file path to directory where you want to
store all your input files.}

\item{target_dir_figure}{Character. Treat this directory as read-only. By 
using the function 'framework::instructions()' the directory will contain 
rendered plots that are specified within your input scripts.}

\item{target_dir_docs}{Character. Treat this directory as read-only. By using
the function 'framework::instructions()' the directory will contain 
rendered documents (pdf, html, ...) of your input scripts.}
}
\description{
This function will prepare implement the given instructions
  (e.g. specify instruction depending on file extension).
}
\note{
This function is part of a family of functions each of which end with
  '_instructions'. The order to call these functions is:
  'prepare_instructions', 'implement_instructions', 'check_instructions',
  'delete_deprecated_instructions', 'execute_instructions' and optionally
  'output_instructions'. There is a wrapper for these functions called
  'instructions'.
}
\seealso{
\code{\link{prepare_instructions}},
  \code{\link{delete_deprecated_instructions}},
  \code{\link{check_instructions}}, \code{\link{instructions}},
  \code{\link{execute_instructions}}, \code{\link{output_instructions}}
}
\author{
Frederik Sachser
}
