% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_srcmake.R, R/insert_srcmake.R
\name{insert_srcmake_addin}
\alias{insert_srcmake_addin}
\title{execute source('make.R')}
\usage{
insert_srcmake_addin()

insert_srcmake_addin()
}
\description{
Call this function as an addin to source the file 'make.R'

Call this function as an addin to print source("make.R") to console.
}

