% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions.R
\name{instructions}
\alias{instructions}
\title{instructions}
\usage{
instructions(source_files, spin_index, cache_index, cache_dir, source_dir,
  data_dir, target_dir_figure, target_dir_docs, target_dir_data)
}
\description{
instructions
}

