% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions_check.R
\name{instructions_check}
\alias{instructions_check}
\title{Check specified instructions}
\usage{
instructions_check(ls_instructions, df_source_files)
}
\arguments{
\item{ls_instructions}{List. A list containing all parameters of a framework project. See ?framework::template_params for a complete list of params needed.}

\item{df_source_files}{data frame. Predefined data frame (output of the function 'instructions_implement').}
}
\value{
updated rds-file (df_source_files.rds) within the cache directory.
}
\description{
Most checks of the given parameters of a framework project are implemented in this function.
}
\note{
This function is part of a family of functions each of which begin with
  'instructions_'. The order to call these functions is:
  'instructions_prepare', 'instructions_implement', 'instructions_check',
  'instructions_execute', 'instructions_symlink' and 'instructions_supplement'.
}
\seealso{
\code{\link{instructions_prepare}}, \code{\link{instructions_implement}},
  \code{\link{instructions_execute}},
  \code{\link{instructions_specify}},
  \code{\link{instructions_symlink}}, \code{\link{instructions_supplement}}
}
\author{
Frederik Sachser
}
