% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions_execute.R
\name{instructions_execute}
\alias{instructions_execute}
\title{Execute your instructions!}
\usage{
instructions_execute(ls_instructions, df_source_files)
}
\arguments{
\item{ls_instructions}{List. A list containing all parameters of a framework project. See ?framework::template_params for a complete list of params needed.}

\item{df_source_files}{data frame. Predefined data frame (output of the function 'instructions_implement').}
}
\value{
rds-files within the cache directory (ls_instructions_last, df_source_files_last, snapshot_data_dir.rds, snapshot_source_dir.rds)
}
\description{
The function will loop through all input_files and call the function instructions_specify.
}
\note{
This function is part of a family of functions each of which begin with
  'instructions_'. The order to call these functions is:
  'instructions_prepare', 'instructions_implement', 'instructions_check',
  'instructions_execute', 'instructions_symlink' and 'instructions_supplement'.
}
\seealso{
\code{\link{instructions_prepare}}, \code{\link{instructions_implement}},
  \code{\link{instructions_check}},
  \code{\link{instructions_specify}},
  \code{\link{instructions_symlink}}, \code{\link{instructions_supplement}}
}
\author{
Frederik Sachser
}
