% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions_prepare.R
\name{instructions_prepare}
\alias{instructions_prepare}
\title{Prepare Instructions}
\usage{
instructions_prepare(ls_instructions)
}
\arguments{
\item{ls_instructions}{List. A list containing all parameters of a framework project. See ?framework::template_params for a complete list of params needed.}
}
\value{
rds-file (ls_instructions.rds) within the cache directory.
}
\description{
This function processes the given ls_instructions.
}
\note{
This function is part of a family of functions each of which begin with
  'instructions_'. The order to call these functions is:
  'instructions_prepare', 'instructions_implement', 'instructions_check',
  'instructions_execute', 'instructions_symlink' and 'instructions_supplement'.
}
\seealso{
\code{\link{instructions_implement}}, \code{\link{instructions_check}},
  \code{\link{instructions_execute}},
  \code{\link{instructions_specify}},
  \code{\link{instructions_symlink}}, \code{\link{instructions_supplement}}
}
\author{
Frederik Sachser
}
