% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions_specify.R
\name{instructions_specify}
\alias{instructions_specify}
\title{Specify instructions!}
\usage{
instructions_specify(input_files, image_cache, instruction, filename_noxt,
  basename_noxt, figure_source, toplvl, rebuild_figures, quiet_processing,
  Rplots_device)
}
\arguments{
\item{input_files}{Character. File path relative to the current working directory. Possible file extensions: Rmd, R and Rnw.}

\item{image_cache}{Character. File path to write an image.}

\item{instruction}{Character. Possible values: 'nothing', 'load', 'source', 'render' and 'knit'.}

\item{filename_noxt}{Character. Complete file path without extension.}

\item{basename_noxt}{Character. Basename of the file without extension.}

\item{figure_source}{Character. File path to write figures.}

\item{toplvl}{Character. Passed to knit_root_dir.}

\item{rebuild_figures}{Logical. Delete figure_source before processing?}

\item{quiet_processing}{Logical. Passed to !echo (if instruction == source) or quiet (if instruction == render or knit).}

\item{Rplots_device}{Function. Specify function that will be passed to a functional to write figures of R-files (if instruction == source). E.g. grDevices::pdf or grDevices::png.}
}
\description{
This function is not meant to be called directly by the user. It
  is integrated into the function 'instructions_execute' for batch processing of input_files.
}
\note{
This function is part of a family of functions each of which begin with
  'instructions_'. The order to call these functions is:
  'instructions_prepare', 'instructions_implement', 'instructions_check',
  'instructions_execute', 'instructions_symlink' and 'instructions_supplement'.
}
\seealso{
\code{\link{instructions_prepare}}, \code{\link{instructions_implement}},
  \code{\link{instructions_check}},
  \code{\link{instructions_execute}},
  \code{\link{instructions_symlink}}, \code{\link{instructions_supplement}}
}
\author{
Frederik Sachser
}
