% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions_supplement.R
\name{instructions_supplement}
\alias{instructions_supplement}
\title{Round out your instructions!}
\usage{
instructions_supplement(ls_instructions)
}
\arguments{
\item{ls_instructions}{List. A list containing all parameters of a framework project. See ?framework::template_params for a complete list of params needed.}
}
\value{
updated rds-file (df_source_files.rds) within the cache directory.
}
\description{
This function will write data frames, session info, log-file, etc and print summaries of these information to the console according to the parameters specified in the file 'params.R'.
}
\note{
This function is part of a family of functions each of which begin with
  'instructions_'. The order to call these functions is:
  'instructions_prepare', 'instructions_implement', 'instructions_check',
  'instructions_execute', 'instructions_symlink' and 'instructions_supplement'.
}
\seealso{
\code{\link{instructions_prepare}}, \code{\link{instructions_implement}},
  \code{\link{instructions_execute}},
  \code{\link{instructions_specify}},
  \code{\link{instructions_symlink}}, \code{\link{instructions_check}}
}
\author{
Frederik Sachser
}
