% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{is_initialized}
\alias{is_initialized}
\title{Check if project is initialized}
\usage{
is_initialized(subdir = NULL)
}
\arguments{
\item{subdir}{Optional subdirectory to check.}
}
\value{
Logical indicating if project is initialized.
}
\description{
Check if project is initialized
}
