% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{ledger_list}
\alias{ledger_list}
\title{List the Project's Cloud Ledger}
\usage{
ledger_list(name = NULL)
}
\arguments{
\item{name}{Optional data catalog name to filter by.}
}
\value{
A data.frame of ledger entries, newest first.
}
\description{
List the Project's Cloud Ledger
}
