% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_framework_templates}
\alias{list_framework_templates}
\title{List available framework templates}
\usage{
list_framework_templates()
}
\description{
List available framework templates
}
\keyword{internal}
