% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{list_projects}
\alias{list_projects}
\title{List all projects in global configuration}
\usage{
list_projects()
}
\value{
Data frame with project information
}
\description{
List all projects in global configuration
}
\keyword{internal}
