% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{load_data_or_cache}
\alias{load_data_or_cache}
\title{Load data with caching}
\usage{
load_data_or_cache(path, expire_after = NULL, refresh = FALSE)
}
\arguments{
\item{path}{Dot notation path to load data (e.g. "source.private.example")}

\item{expire_after}{Optional expiration time in hours (default: from config$options$data$cache_default_expire)}

\item{refresh}{Optional boolean or function that returns boolean to force refresh}
}
\value{
The loaded data, either from cache or file
}
\description{
Load data with caching
}
