% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_entry.R
\name{log_entry}
\alias{log_entry}
\title{write a log entry}
\usage{
log_entry(log_filepath = "log.csv")
}
\arguments{
\item{log_filepath}{Character. Specify relative file path (inlcuding the
extension .csv) to write a log-file.}
}
\value{
A csv file containing current timestamp and nodename.
}
\description{
This function will add a log entry to the specified file.
}
\note{
If the file does not exist it will be written.
}
\seealso{
\code{\link{log_summary}}
}
\author{
Frederik Sachser
}
