% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summary.R
\name{log_summary}
\alias{log_summary}
\title{write a log entry}
\usage{
log_summary(log_filepath = "log.csv")
}
\arguments{
\item{log_filepath}{Character. Specify relative file path (inlcuding the
extension .csv) to write a log-file.}
}
\value{
A list of dataframes including information about duration of work per
  day, per nodename and both.
}
\description{
This function will summarize the information of a specific
  log-file (the output of framework::log_entry()).
}
\note{
The results are not necessarily as intended: It is assumed that you work once a day without breaks. Optionally adjust the logfile as well as the function log_summary to fit your needs.
}
\seealso{
\code{\link{log_entry}}
}
\author{
Frederik Sachser
}
