% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage.R
\name{memory_usage}
\alias{memory_usage}
\title{calculate memory usage}
\usage{
memory_usage(units = "B")
}
\arguments{
\item{units}{Character. Specify dimension of output. See ?utils::object.size for possible values.}
}
\value{
A data frame with the sum of the size of all objects within the global environment and the base environment
}
\description{
Calculate the size of all objects within globalenv and baseenv.
}
\author{
Frederik Sachser
}
