% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages_restore}
\alias{packages_restore}
\title{Restore packages from renv.lock}
\usage{
packages_restore(prompt = FALSE)
}
\arguments{
\item{prompt}{Logical; if TRUE, prompt before restoring}
}
\value{
Invisibly returns TRUE on success
}
\description{
A user-facing wrapper around renv::restore() that ensures renv is enabled.
}
\examples{
\dontrun{
packages_restore()
}
}
