% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cran.R
\name{pkg_cran}
\alias{pkg_cran}
\title{Install and attach packages}
\usage{
pkg_cran(pkg_names, attach = TRUE, ...)
}
\arguments{
\item{pkg_names}{Character. A vector specifying the names of packages you
want to install.}

\item{attach}{Logical. If TRUE the packages will be loaded.}

\item{...}{Further parameters passed to install.packages().}
}
\description{
The function checks whether the specified packages are installed
  or not. Packages that are not installed on your machine are going to be
  installed.
}
\seealso{
\code{\link{pkg_gh}}, \code{\link{install.packages}}
}
\author{
Frederik Sachser
}
