% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_gh.R
\name{pkg_gh}
\alias{pkg_gh}
\title{Install and attach packages}
\usage{
pkg_gh(pkg_names, attach = TRUE, ...)
}
\arguments{
\item{pkg_names}{Character. A vector specifying the names of packages you
want to install.}

\item{attach}{Logical. If TRUE the packages will be loaded.}

\item{...}{Further parameters passed to devtools::install_github().}
}
\description{
The function checks whether the specified packages are installed
  or not. Packages that are not installed on your machine are going to be
  installed.
}
\note{
pkg_names must be specified in the following format: username/repo[/subdir][@ref|#pull]
}
\seealso{
\code{\link{pkg_cran}}, \code{\link[devtools]{install_github}}
}
\author{
Frederik Sachser
}
