% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_instructions.R
\name{prepare_instructions}
\alias{prepare_instructions}
\title{prepare_instructions}
\usage{
prepare_instructions(source_files, spin_index = 0, cache_index = 0,
  cache_dir = ".cache", source_dir = "in/src", data_dir = "in/data",
  target_dir_figure = "out/figure", target_dir_docs = "out/docs")
}
\arguments{
\item{source_files}{eg}

\item{spin_index}{weg}

\item{cache_index}{wesg}

\item{cache_dir}{weg}

\item{source_dir}{weg}

\item{data_dir}{ewg}

\item{target_dir_figure}{ew}

\item{target_dir_docs}{wt}
}
\description{
prepare_instructions
}

