% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_site.R
\name{prepare_site}
\alias{prepare_site}
\title{Prepare and render a local website}
\usage{
prepare_site(Rmd_input, target_dir, index_menu = FALSE, index_name = "Home",
  page_name = basename(getwd()))
}
\arguments{
\item{Rmd_input}{Character. A vector of Rmd-Files that should be placed within
a single directory.}

\item{target_dir}{Character. Path to the output directory of the website.}

\item{index_menu}{Logical. If False there will not be an additional button for 
the main page.}

\item{index_name}{Character. Alternative name for the index (main page).}

\item{page_name}{Character. Name for the homepage.}
}
\value{
A _site.yml and optionally an index.Rmd
}
\description{
The function will prepare given Rmd-files to simplify the process
 of building websites. It should fasten the process of creating webpages by
 using a predefined template and layout. After calling this function you can
 use rmarkdown::render_site() to render a local website.
}
\note{
If there is no "index.Rmd" in input_Rmd a template will be used to 
 create an index_Rmd-file.

Required Version of the package rmarkdown: v0.9.6
}
\author{
Frederik Sachser
}
\seealso{
\code{\link[rmarkdown]{render_site}}
}

