% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_framework.R
\name{project_framework}
\alias{project_framework}
\title{Initialize a framework-project}
\usage{
project_framework(project_dir, rstudio = TRUE, init_git = TRUE,
  init_packrat = FALSE, custom_makeR = NULL, target_makeR = "make.R",
  fun_dir = "R", source_files = c("load.R", "report.Rmd"),
  cache_dir = ".cache", source_dir = "scripts", data_dir = "data",
  target_dir_figure = "out/fig", target_dir_docs = "out/docs",
  target_dir_data = "out/data", devtools_create = TRUE,
  rename_figure = TRUE, rename_docs = TRUE, spin_index = 999,
  cache_index = 999)
}
\arguments{
\item{project_dir}{Character. Specify the path to the directory where you 
want to create a new project. The last directory will be the project 
directory itself. The name of the project will be the name of the project 
directory. E.g. '~/Desktop/MyProject' will create a project directory on 
your desktop.}

\item{rstudio}{Logical. TRUE calls the function Rproj_init.}

\item{init_git}{Logical. TRUE calls the function git_init.}

\item{init_packrat}{Logical. TRUE initializes a packrat repo.}

\item{custom_makeR}{Character. File path to a local make-like R-file. Specify
this option if you want to use a customized version instead of the template
for the file 'make.R'.}

\item{target_makeR}{Character. Target directory of the 'makefile'. Default is
'make.R'. Specify relative file path if you want to use subdirectories.}

\item{fun_dir}{Character. Target directory of functions. Default is 'I-fun'. 
Some functions of the framework-package will be copied to a subdirectory of
this folder. By using the framework template of the file 'make.R' all 
R-Scripts (functions) within this directory will be attached to a 
predefined environment.}

\item{source_files}{Character. Specify a vector of file paths, if you want to
create predefined templates for your analysis (relative to 'source_dir'). 
Use file extensions '.R' or '.Rmd'.}

\item{cache_dir}{Character. Specify file path for the cache directory.}

\item{source_dir}{Character. Specify file path to directory where you want to
store all your input files.}

\item{data_dir}{Character. Specify file path to directory where you want to 
store your data files. This directory will be monitored. When files change 
in this directory the cache will be ignored (when using the framework 
template of the file 'make.R').}

\item{target_dir_figure}{Character. Treat this directory as read-only. By 
using the function 'framework::instructions()' the directory will contain 
rendered plots that are specified within your input scripts.}

\item{target_dir_docs}{Character. Treat this directory as read-only. By using
the function 'framework::instructions()' the directory will contain 
rendered documents (pdf, html, ...) of your input scripts.}

\item{target_dir_data}{Character. Treat this directory as read-only. By using
the function 'framework::write_dataframe()' the directory will contain data
of type 'RData', 'rds' or 'csv'.}

\item{devtools_create}{Logical. Choose TRUE if you want to call 
'devtools::create()' in order to prepare your project as an R-package. You 
might want to choose appropriate paths for other variables (e.g. fun_dir = 
'R', data_dir = inst/extdata, and so on).}

\item{rename_figure}{Logical. If TRUE the rendered figures will be renamed 
and moved into a single directory.}

\item{rename_docs}{Logical. If TRUE the rendered documents (html, pdf, docx) 
will be renamed and moved into a single directory.}

\item{spin_index}{Integer. Can be of length one or a vector. Specify index of
'source_files' that should  be spinned (ignored for file extension 'Rmd'). 
Choose 0 (zero) if you do not want to spin any R-files. Choose 999 if you 
want to spin all R-files.}

\item{cache_index}{Integer. Can be of length one or a vector. Specify index 
of 'source_files' that should  be integrated into the cache. Choose 0 
(zero) if you do not want to use the cache. Choose 999 if you want to use 
the cache for all files.}
}
\description{
This function is probably the only one you need to call by hand.
  Furthermore it is a wrapper for skeleton, Rproj_init and git_init.
  Therefore it is straightforward to create a new project, change the working
  directory, generate basic directories and initialize a git repository.
  Optionally you can initialize a packrat repo.
}
\note{
Creation of the project_dir is recursive.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{Rproj_init}}, \code{\link{git_init}}, 
  \code{\link{skeleton}}
}

