% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{project_list}
\alias{project_list}
\title{List all projects in global configuration}
\usage{
project_list()
}
\value{
Data frame with project information
}
\description{
List all projects in global configuration
}
