\name{project_skeleton}
\alias{project_skeleton}
\title{
Create an empty project skeleton
}
\description{
The function creates directories, a few empty files, a template for a report and a make-like file. 
}
\usage{
project_skeleton()
}
\value{
The output are several files and directories within your working directory.
}
\references{
}
\author{
Frederik Sachser
}
\note{
The function will not overwrite existing files or directories with the same name.
}
\examples{
project_skeleton()
}
\keyword{ project_skeleton }
\keyword{ project_framework }
