% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_execute}
\alias{query_execute}
\title{Execute a database query}
\usage{
query_execute(query, connection_name, ...)
}
\arguments{
\item{query}{SQL query to execute}

\item{connection_name}{Name of the connection in config.yml}

\item{...}{Additional arguments passed to DBI::dbExecute}
}
\value{
Number of rows affected
}
\description{
Executes a query on a database without returning results.
}
