% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminder.R
\name{reminder}
\alias{reminder}
\title{Print a reminder and summary to the console}
\usage{
reminder()
}
\description{
The function will print a summary of the instructions and remind
  you to use git.
}
\author{
Frederik Sachser
}

