% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_delete.R
\name{remove_all_data}
\alias{remove_all_data}
\title{Remove all data values}
\usage{
remove_all_data(delete_files = TRUE)
}
\arguments{
\item{delete_files}{Whether to delete the actual files (default: TRUE)}
}
\description{
Remove all data values
}
