% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_delete.R
\name{remove_data}
\alias{remove_data}
\title{Remove a data value}
\usage{
remove_data(path, delete_file = TRUE)
}
\arguments{
\item{path}{Dot notation path to remove (e.g. "source.private.example")}

\item{delete_file}{Whether to delete the actual file (default: TRUE)}
}
\description{
Remove a data value
}
