% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{remove_init}
\alias{remove_init}
\title{Remove initialization}
\usage{
remove_init(subdir = NULL)
}
\arguments{
\item{subdir}{Optional subdirectory to check.}
}
\value{
Logical indicating if removal was successful.
}
\description{
Removes config.yml to mark project as uninitialized.
WARNING: This will delete your project configuration!
}
