% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{remove_project_from_config}
\alias{remove_project_from_config}
\title{Remove project from global configuration}
\usage{
remove_project_from_config(project_id)
}
\arguments{
\item{project_id}{Project ID to remove}
}
\value{
Invisibly returns NULL
}
\description{
Remove project from global configuration
}
