% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_view.R
\name{restore_framework_view}
\alias{restore_framework_view}
\title{Restore the original View function}
\usage{
restore_framework_view()
}
\value{
Invisibly returns TRUE if successful
}
\description{
This function will restore the original View function if it was overridden.
}
