% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{result_save}
\alias{result_save}
\title{Save a result}
\usage{
result_save(
  name,
  value = NULL,
  type,
  blind = FALSE,
  public = FALSE,
  comment = "",
  file = NULL
)
}
\arguments{
\item{name}{Result name}

\item{value}{The result value to save (will be evaluated if it's an expression)}

\item{type}{Type of result (e.g. "model", "notebook")}

\item{blind}{Whether the result should be blinded (encrypted)}

\item{public}{Whether the result should be public (default: FALSE)}

\item{comment}{Optional description}

\item{file}{Optional path to a file to save (e.g. a QMD notebook)}
}
\description{
Save a result
}
