% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_symlink_dir.R
\name{rm_symlink_dir}
\alias{rm_symlink_dir}
\title{Unlink directory containing symlinks only}
\usage{
rm_symlink_dir(directory)
}
\arguments{
\item{directory}{Character. Specify directory that should be deleted.}
}
\description{
This function will check whether there are only symbolic links within a directory and if so delete the directory.
}
