% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{scratch_clean}
\alias{scratch_clean}
\title{Clean up the scratch directory by deleting all files}
\usage{
scratch_clean()
}
\value{
Invisibly returns NULL. Called for side effect of removing scratch files.
}
\description{
Clean up the scratch directory by deleting all files
}
