% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info.R
\name{session_info}
\alias{session_info}
\title{Write a file with current session info}
\usage{
session_info(session_info_filepath = "meta/session_info.txt")
}
\arguments{
\item{session_info_filepath}{Character. File path to the target.}
}
\description{
The function writes the current timestamp and session info to a
  file.
}
\note{
Directories will be created recursively.
}
\author{
Frederik Sachser
}

