% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_srcmake.R
\name{srcmake_addin}
\alias{srcmake_addin}
\title{Execute source('make.R')}
\usage{
srcmake_addin()
}
\description{
This addin only works when using RStudio. Call this function as an addin to
source the file 'make.R'. Optionally it is possible to assign a custom
keybinding (see RStudio/Tools/Modify keyboard shortcuts... and search for
"Execute source('make.R')").
}
