% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_git.R
\name{summary_git}
\alias{summary_git}
\title{Summarize a git repository}
\usage{
summary_git(git_repo, filepath_git_summary)
}
\arguments{
\item{git_repo}{Character. File path to your repository.}

\item{filepath_git_summary}{Character. File path to the target.}
}
\description{
This function is an alias for git2r::summary. It will print the output to the console as well as redirect it to a file.
}
\author{
Frederik Sachser
}
