% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_instructions.R
\name{summary_instructions}
\alias{summary_instructions}
\title{Print a reminder and summary to the console}
\usage{
summary_instructions()
}
\description{
The function will print a summary of the instructions.
}
\author{
Frederik Sachser
}

