% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_warnings.R
\name{summary_warnings}
\alias{summary_warnings}
\title{Summarize warnings in a framework project}
\usage{
summary_warnings(filepath_warnings = "meta/warnings.Rout")
}
\arguments{
\item{filepath_warnings}{Character. File path to the file containing collected warnings}
}
\description{
Extract lines from file and print to console (coloured warnings and messages).
}
\seealso{
\code{\link{write_warnings}}
}
\author{
Frederik Sachser
}
