% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_html.R
\name{template_html}
\alias{template_html}
\title{Create a template script-file}
\usage{
template_html(file, Author = Sys.info()["effective_user"],
  Date = "`r Sys.Date()`", open = TRUE)
}
\arguments{
\item{file}{Character. Specify the path to save the new file. Use relative 
file paths and specify the file extension; e.g. 
'in/src/new_dir/myfile.Rmd'. or 'in/in/src/new_dir/myfile.R'. If the
extension is neither .R nor .Rmd (e.g. no extension) an .Rmd-extension will
be added by default.}

\item{Author}{Character. Optionally customize the name of the Author (used
for the YAML header). Default is the effective user of the system info.}

\item{Date}{Character. Optionally customize the date (used for the YAML
header). Default is the current Date (format YYYY-MM-DD).}

\item{open}{Logical. If TRUE the file will be opened (via `file.edit``).}
}
\description{
This function will create an R or Rmd-file (depending on the
  file extension of the input) including some predefined lines. By using this
  template for Rmd-files within a framework-project it is straightforward to
  save all plots you will specify within the script as pdf and png version
  (using rmarkdown::render). The function is designed to save some time
  writing the same input again and again (e.g. Author, date and other
  things).
}
\note{
Missing directories will be created recursively.

It is not possible to overwrite existing files.

Other YAML header options will be choosen automatically. Edit the
  resulting file to customize the YAML header.
}
\author{
Frederik Sachser
}

