% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_make.R
\name{template_make}
\alias{template_make}
\title{template_make}
\usage{
template_make(target_makeR = "make.R", fun_dir = "in/fun",
  source_files = c("load.R", "report.Rmd"), cache_dir = ".cache",
  source_dir = "in/src", data_dir = "in/data",
  target_dir_figure = "out/figure", target_dir_docs = "out/docs",
  target_dir_data = "out/data")
}
\description{
template_make
}

