% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedir.R
\name{treedir}
\alias{treedir}
\title{Tree-view a directory structure}
\usage{
treedir(tree_directory = getwd(), tree_target, include_hidden = FALSE)
}
\arguments{
\item{tree_directory}{Character. Specify a path to a directory. default: current working directory.}

\item{tree_target}{Character. Optionally specify a file path to write the output to a file. default: stdout.}

\item{include_hidden}{Logical. If set to false: all hidden directories and files will be omitted.}
}
\description{
This function will summarize a given directory structure as a tree.
}
\author{
Frederik Sachser
}
