% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_view.R
\name{use_framework_view}
\alias{use_framework_view}
\title{Override the default View function with the Framework View version}
\usage{
use_framework_view()
}
\value{
Invisibly returns TRUE if successful
}
\description{
This function will replace the default View function with our Framework View version.
It should be called when loading the framework.
}
