% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{website}
\alias{website}
\title{Prepare and render a local website}
\usage{
website(Rmd_input, target_dir, ...)
}
\arguments{
\item{Rmd_input}{Character. A vector of Rmd-Files, that should be parts of the
website.}

\item{target_dir}{Character. Path to the output directory.}

\item{...}{further options for the function rmarkdown::render_site}
}
\value{
there will be two output directories: Rmd and html.
}
\description{
The function will copy the Rmd_input, write a _site.yml and use 
 rmarkdown to render a website on your local machine. It is a wrapper for the
 function rmarkdown::render_site. The function should fasten the process of
 creating webpages by using a predefined template and layout.
}
\note{
If there is no "index.Rmd" in input_Rmd a template will be used to
 create an index_Rmd-file.

Required Version of the package rmarkdown: v0.9.6
}
\author{
Frederik Sachser
}
\references{
JJ Allaire, Joe Cheng, Yihui Xie, Jonathan McPherson, Winston
 Chang, Jeff Allen, Hadley Wickham, Aron Atkins and Rob Hyndman (2016).
 rmarkdown: Dynamic Documents for R. 
 R package version 1.0. 
 https://CRAN.R-project.org/package=rmarkdown
}
\seealso{
\code{\link[rmarkdown]{render_site}}
}

