% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_framework_template}
\alias{write_framework_template}
\title{Overwrite a Framework template with new contents}
\usage{
write_framework_template(name, contents)
}
\arguments{
\item{name}{Template identifier (e.g., "notebook", "gitignore", "ai_claude")}

\item{contents}{Character string to write to the template file.}
}
\value{
Invisibly returns the file path of the written template.
}
\description{
Overwrite a Framework template with new contents
}
