% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_index_Rmd.R
\name{write_index_Rmd}
\alias{write_index_Rmd}
\title{dsfrrgroodsgh}
\usage{
write_index_Rmd(target_index_Rmd = ".cache/website/index.Rmd")
}
\arguments{
\item{target_index_Rmd}{some text}
}
\description{
aswgarsg swegoo
}
\author{
Frederik Sachser
}

