% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{write a log entry}
\usage{
write_log(filepath_log = "meta/log.csv", summarize_log = TRUE)
}
\arguments{
\item{filepath_log}{Character. Specify file path to write the log file.}

\item{summarize_log}{Logical. Should a summary of the log file be printed to console? Depends on dplyr.}
}
\value{
A csv file containing current timestamp and nodename.
}
\description{
This function will add a log entry to the specified file.
}
\note{
If the file does not exist it will be written.
}
\seealso{
\code{\link{summary_log}}
}
\author{
Frederik Sachser
}
