% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_session_info.R
\name{write_session_info}
\alias{write_session_info}
\title{Write a file with current session info}
\usage{
write_session_info(filepath_session_info = "meta/session_info.txt",
  summarize_session_info = FALSE)
}
\arguments{
\item{filepath_session_info}{Character. Specify file path to write the session info file.}

\item{summarize_session_info}{Logical. Should session info be printed to console?}
}
\description{
The function writes the current timestamp and session info to a
  file.
}
\note{
Directories will be created recursively.
}
\author{
Frederik Sachser
}
