% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tree.R
\name{write_tree}
\alias{write_tree}
\title{Tree-view a directory structure}
\usage{
write_tree(tree_directory = getwd(), filepath_tree,
  include_hidden_tree = FALSE, summarize_tree = TRUE)
}
\arguments{
\item{tree_directory}{Character. Specify input directory as top level for the tree.}

\item{filepath_tree}{Character. Specify file path to write the tree file.}

\item{include_hidden_tree}{Logical. Include/Exclude hidden directories and files to build the tree.}

\item{summarize_tree}{Logical. Should the directory structure (tree) be printed to console?}
}
\description{
This function will summarize a given directory structure as a tree.
}
\author{
Frederik Sachser
}
