% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_warnings.R
\name{write_warnings}
\alias{write_warnings}
\title{Append _warnings.Rout of current input files}
\usage{
write_warnings(filepath_warnings, summarize_warnings = FALSE)
}
\arguments{
\item{filepath_warnings}{Character. Specify file path to write the warnings file.}

\item{summarize_warnings}{Logical. Should filepath_warnings be printed to console?}
}
\description{
This function will put together all '_warnings.Rout'-files of current input files.
}
\note{
The function assumes you are working in a framework project and will otherwise fail.
}
\seealso{
\code{\link{summary_warnings}}
}
\author{
Frederik Sachser
}
