% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xview.R
\name{xview}
\alias{xview}
\title{External View}
\usage{
xview(x, file = "xview.xlsx", ...)
}
\arguments{
\item{x}{Object to export (e.g. a data.frame)}

\item{file}{Specify filename. Directories will not be created.}

\item{...}{Further params passed to rio::export (e.g. overwrite, row.names, etc.)}
}
\description{
This function will export an object to a file (using package "rio"). A system command is invoked to open the file afterwards.
}
\seealso{
\code{\link[rio]{export}}
}
