% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{erhth}
\usage{
backup(target_dir = "project_subdir", source_dir = file.path(getwd()),
  overwrite = TRUE, exclude_directories = FALSE, exclude_files = FALSE)
}
\description{
srg argrag
}

