% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_read.R
\name{cache_get}
\alias{cache_get}
\title{Get a cached value}
\usage{
cache_get(name)
}
\arguments{
\item{name}{The cache name}
}
\value{
The cached value, or NULL if not found or hash mismatch
}
\description{
Get a cached value
}
