% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{connection_update}
\alias{connection_update}
\title{Update a record in a table}
\usage{
connection_update(conn, table_name, id, values, auto_timestamps = TRUE)
}
\arguments{
\item{conn}{Database connection}

\item{table_name}{Name of the table}

\item{id}{The ID of the record to update}

\item{values}{Named list of column-value pairs to update}

\item{auto_timestamps}{Whether to automatically set updated_at (default: TRUE)}
}
\value{
Number of rows affected
}
\description{
Updates an existing record in a table with automatic timestamp handling.
If the table has an updated_at column, it will be set automatically.
}
\examples{
\dontrun{
conn <- connection_get("postgres")
connection_update(conn, "users", 1, list(age = 31, name = "Alice Updated"))
DBI::dbDisconnect(conn)
}

}
