% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_write.R
\name{data_save}
\alias{data_save}
\title{Save data using dot notation path}
\usage{
data_save(
  data,
  path,
  type = "csv",
  delimiter = "comma",
  locked = TRUE,
  encrypted = FALSE
)
}
\arguments{
\item{data}{Data frame to save}

\item{path}{Dot notation path to save data (e.g. "source.private.example")}

\item{type}{Type of data file ("csv" or "rds")}

\item{delimiter}{Delimiter for CSV files ("comma", "tab", "semicolon", "space")}

\item{locked}{Whether the file should be locked after saving}

\item{encrypted}{Whether the file should be encrypted}
}
\description{
Save data using dot notation path
}
