% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{data_spec_get}
\alias{data_spec_get}
\title{Get data specification (DEPRECATED)}
\usage{
data_spec_get(path)
}
\arguments{
\item{path}{Dot notation path or file path}
}
\value{
A list with data specification
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been renamed to \code{\link[=data_info]{data_info()}}. Please use that instead.
}
\keyword{internal}
