% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_deprecated_instructions.R
\name{delete_deprecated_instructions}
\alias{delete_deprecated_instructions}
\title{Keep your output clean!}
\usage{
delete_deprecated_instructions(source_files, cache_dir, data_dir,
  df_source_files)
}
\arguments{
\item{source_files}{Character. Specify a vector of file paths, if you want to
create predefined templates for your analysis (relative to top-level). 
Use file extensions '.R' or '.Rmd'.}

\item{cache_dir}{Character. Specify file path for the cache directory.}

\item{data_dir}{Character. Specify file path to directory where you want to 
store your data files. This directory will be monitored. When files change 
in this directory the cache will be ignored (when using the framework 
template of the file 'make.R').}

\item{df_source_files}{Object of class data.frame (i.e. output of the function framework::implement_instructions).}
}
\description{
This function will delete all deprecated files frem your last
  session.
}
\note{
This function was not designed to be run separately. Use framework::instructions() instead.
}
\seealso{
\code{\link{prepare_instructions}},
  \code{\link{implement_instructions}}, \code{\link{check_instructions}},
  \code{\link{instructions}}, \code{\link{execute_instructions}},
  \code{\link{output_instructions}}
}
\author{
Frederik Sachser
}
