% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_deprecated_instructions.R
\name{delete_deprecated_instructions}
\alias{delete_deprecated_instructions}
\title{Keep your output clean!}
\usage{
delete_deprecated_instructions(source_files, cache_dir, data_dir,
  df_source_files)
}
\arguments{
\item{source_files}{Character. Specify a vector of file paths, if you want to
create predefined templates for your analysis (relative to top-level). 
Use file extensions '.R' or '.Rmd'.}

\item{cache_dir}{Character. Specify file path for the cache directory.}

\item{data_dir}{Character. Specify file path to directory where you want to 
store your data files. This directory will be monitored. When files change 
in this directory the cache will be ignored (when using the framework 
template of the file 'make.R').}
}
\description{
This function will delete all deprecated files frem your last
  session.
}
\note{
This function is part of a family of functions each of which end with
  '_instructions'. The order to call these functions is:
  'prepare_instructions', 'implement_instructions', 'check_instructions',
  'delete_deprecated_instructions', 'execute_instructions' and optionally
  'output_instructions'. There is a wrapper for these functions called
  'instructions'.
}
\seealso{
\code{\link{prepare_instructions}},
  \code{\link{implement_instructions}}, \code{\link{check_instructions}},
  \code{\link{instructions}}, \code{\link{execute_instructions}},
  \code{\link{output_instructions}}
}
\author{
Frederik Sachser
}
