% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{.commit_after_scaffold}
\alias{.commit_after_scaffold}
\title{Create initial commit after first successful scaffold}
\usage{
.commit_after_scaffold()
}
\description{
Create initial commit after first successful scaffold
}
\keyword{internal}
