% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.create_config_file}
\alias{.create_config_file}
\title{Create config.yml from template}
\usage{
.create_config_file(type = "analysis", subdir = NULL)
}
\description{
Create config.yml from template
}
\keyword{internal}
