% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption_core.R
\name{.encrypt_with_password}
\alias{.encrypt_with_password}
\title{Encrypt data with password}
\usage{
.encrypt_with_password(data, password)
}
\arguments{
\item{data}{Raw vector to encrypt}

\item{password}{Character string password}
}
\value{
Raw vector with header and encrypted data
}
\description{
Encrypts raw data using password-derived key.
}
\details{
File format:
\link{MAGIC (6)} \link{SALT (32)} \link{NONCE (24)} \link{RESERVED (1)} \link{CIPHERTEXT (variable)}
}
\keyword{internal}
