% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworkdb.R
\name{.get_db_connection}
\alias{.get_db_connection}
\title{Get a connection to the framework database}
\usage{
.get_db_connection()
}
\description{
Get a connection to the framework database
}
\keyword{internal}
