% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_db.R
\name{.get_metadata}
\alias{.get_metadata}
\title{Get a metadata value}
\usage{
.get_metadata(key)
}
\arguments{
\item{key}{The metadata key}
}
\value{
The metadata value, or NULL if not found
}
\description{
Get a metadata value
}
\keyword{internal}
