% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.init_standard}
\alias{.init_standard}
\title{Standard initialization process (shared by both paths)}
\usage{
.init_standard(project_name, project_structure, lintr, styler, subdir, force)
}
\description{
Standard initialization process (shared by both paths)
}
\keyword{internal}
