% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.init_standard}
\alias{.init_standard}
\title{Standard initialization process (shared by both paths)}
\usage{
.init_standard(
  project_name,
  type,
  lintr,
  author_name = NULL,
  author_email = NULL,
  author_affiliation = NULL,
  default_notebook_format = NULL,
  subdir,
  force,
  use_git = TRUE
)
}
\description{
Standard initialization process (shared by both paths)
}
\keyword{internal}
