% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_write.R
\name{.set_cache}
\alias{.set_cache}
\title{Set a cache value}
\usage{
.set_cache(name, value, file = NULL, expire_after = NULL)
}
\arguments{
\item{name}{The cache name}

\item{value}{The value to cache}

\item{file}{Optional file path to store the cache (default: {config$options$data$cache_dir}/{name}.rds)}

\item{expire_after}{Optional expiration time in hours (default: from config$options$data$cache_default_expire)}
}
\description{
Set a cache value
}
\keyword{internal}
