% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_instructions.R
\name{execute_instructions}
\alias{execute_instructions}
\title{Execute your instructions!}
\usage{
execute_instructions(cache_dir = ".cache")
}
\arguments{
\item{cache_dir}{Character. Specify file path for the cache directory.}
}
\description{
The function will execute the instructions (source/render
  specified files etc). Among other things it is a wrapper for the function
  'specify_instructions'.
}
\note{
This function is part of a family of functions each of which end with
  '_instructions'. The order to call these functions is:
  'prepare_instructions', 'implement_instructions', 'check_instructions',
  'delete_deprecated_instructions', 'execute_instructions' and optionally
  'output_instructions'. There is a wrapper for these functions called
  'instructions'.
}
\seealso{
\code{\link{prepare_instructions}},
  \code{\link{implement_instructions}}, \code{\link{check_instructions}},
  \code{\link{delete_deprecated_instructions}}, \code{\link{instructions}},
  \code{\link{output_instructions}}, \code{\link{specify_instructions}}
}
\author{
Frederik Sachser
}
