% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{Alias for file.edit()}
\usage{
fe(...)
}
\arguments{
\item{...}{Arguments passed to file.edit().}
}
\description{
This function is just a (shorter) assignment for file.edit().
}
\seealso{
\code{\link[utils]{file.edit}}
}
