% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-package.R
\docType{package}
\name{framework-package}
\alias{framework}
\alias{framework-package}
\title{framework: A Framework for Data Management}
\description{
A framework for managing data files, configurations, and database connections.
}
\author{
\strong{Maintainer}: Erik Westlund \email{ewestlund@jhu.edu} [copyright holder]

}
\keyword{internal}
