% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_view.R
\name{framework_view}
\alias{framework_view}
\title{Framework View function that opens a DataTable in the browser using DT package}
\usage{
framework_view(x, title = NULL, max_rows = 5000)
}
\arguments{
\item{x}{The data to view}

\item{title}{Optional title for the view}

\item{max_rows}{Maximum number of rows to display for data frames (default: 1000)}
}
\value{
Opens a browser window with the data in a DataTable
}
\description{
Framework View function that opens a DataTable in the browser using DT package
}
