% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_connection}
\alias{get_connection}
\title{Get a database connection from config}
\usage{
get_connection(name)
}
\arguments{
\item{name}{Name of the connection in config.yml (e.g. "postgres")}
}
\value{
A database connection object
}
\description{
Gets a database connection based on the connection name in config.yml.
}
