% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_instructions.R
\name{implement_instructions}
\alias{implement_instructions}
\title{Implement specified instructions}
\usage{
implement_instructions(cache_dir = ".cache")
}
\arguments{
\item{cache_dir}{Character. Specify file path for the cache directory.}
}
\description{
This function will prepare implement the given instructions
  (e.g. specify instruction depending on file extension).
}
\note{
This function is part of a family of functions each of which end with 
  '_instructions'. The order to call these functions is: 
  'prepare_instructions', 'implement_instructions', 'check_instructions', 
  'delete_deprecated_instructions', 'execute_instructions' and optionally 
  'output_instructions'. There is a wrapper for these functions called 
  'instructions'.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{prepare_instructions}}, 
  \code{\link{delete_deprecated_instructions}},
  \code{\link{check_instructions}}, \code{\link{instructions}},
  \code{\link{execute_instructions}}, \code{\link{output_instructions}}
}

