% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the framework}
\usage{
init(
  project_name = NULL,
  project_structure = "default",
  lintr = "default",
  styler = "default",
  subdir = NULL,
  force = FALSE
)
}
\arguments{
\item{project_name}{The name of the project (used for .Rproj file).}

\item{project_structure}{The project structure to use.}

\item{lintr}{The lintr style to use.}

\item{styler}{The styler style to use.}

\item{subdir}{Optional subdirectory to copy files into. If provided, {subdir} in config files will be replaced with subdir/.}

\item{force}{If TRUE, will reinitialize even if project is already initialized.}
}
\description{
This function initializes the framework for a new project.
It sets up the necessary files in the current directory.
}
