% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_srcmake.R
\name{insert_srcmake_addin}
\alias{insert_srcmake_addin}
\title{Insert text to source make.R}
\usage{
insert_srcmake_addin()
}
\description{
Call this function as an addin to insert preformatted text at the cursor position.
}

