% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_srcmake.R
\name{insert_srcmake_addin}
\alias{insert_srcmake_addin}
\title{Insert source('make.R') to console}
\usage{
insert_srcmake_addin()
}
\description{
Call this function as an addin to print source("make.R") to console.
}

