% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_srcmake.R
\name{insert_srcmake_addin}
\alias{insert_srcmake_addin}
\title{Insert source('make.R') to console}
\usage{
insert_srcmake_addin()
}
\description{
This addin only works when using RStudio. Call this function as an addin to
print source("make.R") to console. Optionally it is possible to assign a custom
keybinding (see RStudio/Tools/Modify keyboard shortcuts... and search for
"Execute source('make.R')").
}
