% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_framework_templates}
\alias{list_framework_templates}
\title{Return metadata about available templates}
\usage{
list_framework_templates()
}
\description{
Return metadata about available templates
}
\keyword{List}
\keyword{available}
\keyword{framework}
\keyword{internal}
\keyword{templates}
