% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_instructions.R
\name{prepare_instructions}
\alias{prepare_instructions}
\title{prepare_instructions}
\usage{
prepare_instructions(source_files, spin_index = 0, cache_index = 0,
  cache_dir = ".cache", source_dir = "in/src", data_dir = "in/data",
  target_dir_figure = "out/figure", target_dir_docs = "out/docs",
  rename_figure = TRUE, rename_docs = TRUE, knitr_cache = TRUE)
}
\arguments{
\item{source_files}{Character. Specify a vector of file paths, if you want to
create predefined templates for your analysis (relative to 'source_dir'). 
Use file extensions '.R' or '.Rmd'.}

\item{spin_index}{Integer. Can be of length one or a vector. Specify index of
'source_files' that should  be spinned (ignored for file extension 'Rmd'). 
Choose 0 (zero) if you do not want to spin any R-files. Choose 999 if you 
want to spin all R-files.}

\item{cache_index}{Integer. Can be of length one or a vector. Specify index 
of 'source_files' that should  be integrated into the cache. Choose 0 
(zero) if you do not want to use the cache. Choose 999 if you want to use 
the cache for all files.}

\item{cache_dir}{Character. Specify file path for the cache directory.}

\item{source_dir}{Character. Specify file path to directory where you want to
store all your input files.}

\item{data_dir}{Character. Specify file path to directory where you want to 
store your data files. This directory will be monitored. When files change 
in this directory the cache will be ignored (when using the framework 
template of the file 'make.R').}

\item{target_dir_figure}{Character. Treat this directory as read-only. By 
using the function 'framework::instructions()' the directory will contain 
rendered plots that are specified within your input scripts.}

\item{target_dir_docs}{Character. Treat this directory as read-only. By using
the function 'framework::instructions()' the directory will contain 
rendered documents (pdf, html, ...) of your input scripts.}

\item{rename_figure}{Logical. If TRUE the rendered figures will be renamed 
and moved into a single directory.}

\item{rename_docs}{Logical. If TRUE the rendered documents (html, pdf, docx) 
will be renamed and moved into a single directory.}

\item{knitr_cache}{Logical. If you want to use the package 'knitr' to cache 
chunks of a file you should additionally specify knitr_cache = TRUE within 
the function 'instructions'. By choosing this option it is not possible to 
use a different target for your figures (target_dir_figure = NULL).}
}
\value{
RData file within the cache directory containing all information of
  the given instructions.
}
\description{
This function processes the given instructions.
}
\note{
This function is part of a family of functions each of which end with
  '_instructions'. The order to call these functions is:
  'prepare_instructions', 'implement_instructions', 'check_instructions',
  'delete_deprecated_instructions', 'execute_instructions' and optionally
  'output_instructions'. There is a wrapper for these functions called
  'instructions'.
}
\seealso{
\code{\link{instructions}}, \code{\link{implement_instructions}},
  \code{\link{check_instructions}},
  \code{\link{delete_deprecated_instructions}},
  \code{\link{execute_instructions}}, \code{\link{output_instructions}}
}
\author{
Frederik Sachser
}
