% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_framework.R
\name{project_framework}
\alias{project_framework}
\title{dsfrrgrggdsgh}
\usage{
project_framework(dirname, init_Rproj = TRUE, init_git = TRUE,
  init_packrat = FALSE)
}
\description{
aswgarsg sweggda
}
\author{
Frederik Sachser
}

