\name{project_framework}
\alias{project_framework}
\title{
Create an empty project framework
}
\description{
The function should be called at the very beginning of a project. It creates directories, a few empty files, a report and a make-like file. By using this framework it is straighforward to standardize project structures.
}
\usage{
project_framework()
}
\value{
The output are several files and directories within your working directory.
}
\author{
Frederik Sachser
}
\note{
The files and directories will not overwrite existing files or directories.
}
\examples{
project_framework()
}
\keyword{ project_framework }
