\name{skeleton}
\alias{skeleton}
\title{
Create an empty project skeleton
}
\description{
The function creates a predefined directory structure and some files to enhance project organization.
}
\usage{
skeleton()
}
\value{
The output are several files and directories within your working directory.
}
\references{
}
\author{
Frederik Sachser
}
\note{
The function will not overwrite existing files or directories with the same name.
}
\examples{
skeleton()
}
\keyword{ skeleton }
\keyword{ project_framework }
