% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_get}
\alias{query_get}
\title{Get data from a database query}
\usage{
query_get(query, connection_name, ...)
}
\arguments{
\item{query}{SQL query to execute}

\item{connection_name}{Name of the connection in config.yml}

\item{...}{Additional arguments passed to DBI::dbGetQuery}
}
\value{
A data frame with the query results
}
\description{
Gets data from a database using a query and connection name. The connection
is created, used, and automatically closed. For better performance with
multiple queries, enable connection pooling via \code{pool: true} in your
connection configuration.
}
\examples{
\dontrun{
# Default: creates new connection, auto-disconnects
users <- query_get("SELECT * FROM users", "my_db")

# To use pooling, configure in config.yml:
# connections:
#   my_db:
#     driver: postgres
#     pool: true
}

}
