% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read project configuration}
\usage{
read_config()
}
\value{
The configuration as a list
}
\description{
Reads the project configuration from config.yml, evaluating any expressions.
}
