% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{result_get}
\alias{result_get}
\title{Get a result}
\usage{
result_get(name)
}
\arguments{
\item{name}{Result name}
}
\value{
The result value, or NULL if not found or hash mismatch
}
\description{
Get a result
}
