% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rproj_init.R
\name{Rproj_init}
\alias{Rproj_init}
\title{Initialize an RStudio-project}
\usage{
Rproj_init()
}
\description{
Creates an .Rproj-file within your current working directory.
}
\note{
It is not possible to overwrite an existing Rproj.
}
\author{
Frederik Sachser
}

