% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{Create a framework-project skeleton}
\usage{
skeleton(custom_makeR = NULL, target_makeR = "make.R", fun_dir = "in/fun",
  source_files = c("load.R", "report.Rmd"), cache_dir = ".cache",
  source_dir = "in/src", data_dir = "in/data",
  target_dir_figure = "out/figure", target_dir_docs = "out/docs",
  target_dir_data = "out/data")
}
\arguments{
\item{custom_makeR}{Character. File path to a local make-like R-file. Specify
this option if you want to use a customized version instead of the template
for the file 'make.R'.}
}
\value{
The output are several files and directories within your working 
  directory. The structure contains a how-to-guide.txt, README.md, 
  directories for input and output, some file-templates and several functions
  for proper functioning of the project.
}
\description{
The function creates a predefined directory structure and some 
  files to enhance project organization.
}
\note{
There are no parameters to specify.

There is a wrapper-function to call this function. You might want to 
  use project_framework() instead.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{project_framework}}
}

