% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_skeleton.R, R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{Create a framework-project skeleton}
\usage{
skeleton()

skeleton()
}
\value{
The output are several files and directories within your working
  directory. The structure contains a how-to-guide.txt, README.md,
  directories for input and output, some file-templates and several functions
  for proper functioning of the project.

The output are several files and directories within your working
  directory. The structure contains a how-to-guide.txt, README.md,
  directories for input and output, some file-templates and several functions
  for proper functioning of the project.
}
\description{
The function creates a predefined directory structure and some
  files to enhance project organization.

The function creates a predefined directory structure and some
  files to enhance project organization.
}
\note{
There are no parameters to specify.

There is a wrapper-function to call this function. You might want to
  use project_framework() instead.

There are no parameters to specify.

There is a wrapper-function to call this function. You might want to
  use project_framework() instead.
}
\author{
Frederik Sachser

Frederik Sachser
}
\seealso{
\code{\link{project_framework}}

\code{\link{project_framework}}
}

