% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_instructions.R
\name{specify_instructions}
\alias{specify_instructions}
\title{dsfrrgrjjdsgh}
\usage{
specify_instructions(filename_in, basename_in, filename_image, filename_mtime,
  filename_Rmd, instruction, filename_dot)
}
\description{
aswgarsg swegjj
}
\author{
Frederik Sachser
}

