% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_instructions.R
\name{specify_instructions}
\alias{specify_instructions}
\title{Specify instructions!}
\usage{
specify_instructions(target_dir_docs, target_dir_figure, source_dir, filename,
  image_cache, instruction, docs_out, filename_noxt, basename_noxt,
  temp_docs_out, figure_source, figure_out, use_spin, file_ext, knitr_cache)
}
\arguments{
\item{target_dir_docs}{Character. Treat this directory as read-only. By using
the function 'framework::instructions()' the directory will contain 
rendered documents (pdf, html, ...) of your input scripts.}

\item{target_dir_figure}{Character. Treat this directory as read-only. By 
using the function 'framework::instructions()' the directory will contain 
rendered plots that are specified within your input scripts.}

\item{source_dir}{Character. Specify file path to directory where you want to
store all your input files.}

\item{filename}{Column output from framework::implement_instructions.}

\item{image_cache}{Column output from framework::implement_instructions.}

\item{instruction}{Column output from framework::implement_instructions.}

\item{docs_out}{Column output from framework::implement_instructions.}

\item{filename_noxt}{Column output from framework::implement_instructions.}

\item{basename_noxt}{Column output from framework::implement_instructions.}

\item{temp_docs_out}{Column output from framework::implement_instructions.}

\item{figure_source}{Column output from framework::implement_instructions.}

\item{figure_out}{Column output from framework::implement_instructions.}

\item{use_spin}{Column output from framework::implement_instructions.}

\item{file_ext}{Column output from framework::implement_instructions.}

\item{knitr_cache}{Logical. If you want to use the package 'knitr' to cache 
chunks of a file you should additionally specify knitr_cache = TRUE within 
the function 'instructions'. By choosing this option it is not possible to 
use a different target for your figures (target_dir_figure = NULL).}
}
\description{
This function will source/render/knit input files depending on the parameters.
}
\note{
This function was not designed to be run separately. Use framework::instructions() instead.
}
\seealso{
\code{\link{prepare_instructions}},
  \code{\link{implement_instructions}}, \code{\link{check_instructions}},
  \code{\link{instructions}}, \code{\link{execute_instructions}},
  \code{\link{output_instructions}}
}
\author{
Frederik Sachser
}
