% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_git.R
\name{summary_git}
\alias{summary_git}
\title{Wrapper for git2r::summary}
\usage{
summary_git(git_repo)
}
\arguments{
\item{git_repo}{Character. Directory of your git repository.}
}
\description{
Wrapper for git2r::summary
}
\seealso{
\code{\link[utils]{installed.packages}}, \code{\link[git2r]{repository}}
}
