% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_instructions.R
\name{summary_instructions}
\alias{summary_instructions}
\title{Summarize what has been done!}
\usage{
summary_instructions(cache_dir)
}
\arguments{
\item{cache_dir}{Character. Directory for the cache.}
}
\description{
This function reads information from your cache and prints the
  filenames as well as related instructions. Therefore you can track the
  executed instructions. The information corresponds to the point in time
  when you run the function 'execute_instructions'.
}
\author{
Frederik Sachser
}
