% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_make.R
\name{template_make}
\alias{template_make}
\title{Write a predefined file make.R}
\usage{
template_make()
}
\value{
A file named 'make.R'.
}
\description{
Aside from scripts for analysis each framework project contain 
  two important files: 'make.R' and 'params.R'. All parameters should be 
  specified in params.R. These values will be passed to make.R
}
\seealso{
\code{\link{template_params}}
}
\author{
Frederik Sachser
}
