% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_R.R
\name{template_R}
\alias{template_R}
\title{Create a template R-file}
\usage{
template_R(file, Author = Sys.info()["effective_user"],
  Date = "`r Sys.Date()`")
}
\arguments{
\item{file}{Character. Specify the path to save the new file. Use relative file paths and specify the file extension; e.g. 'in/src/new_dir/myfile.R'.}

\item{Author}{Character. Optionally customize the name of the Author (used for the YAML header). Default is the effective user of the system info.}

\item{Date}{Character. Optionally customize the date (used for the YAML header). Default is the current Date (format YYYY-MM-DD).}
}
\description{
Create a template R-file
}
\note{
Missing directories will be created recursively.

It is not possible to overwrite existing files.

The file contains a YAML header and Roxygen comment prefixes. It is possible to source these file as well as to spin it (using knitr).

Other YAML header options will be choosen automatically. Edit the resulting file to customize the YAML header.
}
\author{
Frederik Sachser
}

