% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_Rmd.R
\name{template_Rmd}
\alias{template_Rmd}
\title{Create a template Rmd-file}
\usage{
template_Rmd(file, Author = Sys.info()["effective_user"],
  Date = "`r Sys.Date()`")
}
\arguments{
\item{file}{Character. Specify the path to save the new file. Use relative file paths and specify the file extension; e.g. 'in/src/new_dir/myfile.Rmd'.}

\item{Author}{Character. Optionally customize the name of the Author (used for the YAML header). Default is the effective user of the system info.}

\item{Date}{Character. Optionally customize the date (used for the YAML header). Default is the current Date (format YYYY-MM-DD).}
}
\description{
Create a template Rmd-file
}
\note{
Missing directories will be created recursively.

It is not possible to overwrite existing files.

Other YAML header options will be choosen automatically. Edit the resulting file to customize the YAML header.
}
\author{
Frederik Sachser
}

