% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframe.R
\name{write_dataframe}
\alias{write_dataframe}
\title{dsfrrgrdsghrr}
\usage{
write_dataframe(listofdf = "GlobalEnv", target_dir = "out/data/",
  file_format = "csv")
}
\description{
aswgarsg swegrr
}
\author{
Frederik Sachser
}

