% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframe.R
\name{write_dataframe}
\alias{write_dataframe}
\title{write_dataframe}
\usage{
write_dataframe(listofdf = "GlobalEnv", target_dir_data = "out/data",
  file_format = "csv")
}
\description{
write all dataframes
}

