% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_yaml}
\alias{write_yaml}
\title{dsfrrgrqqdsgh}
\usage{
write_yaml(target_dir_yaml = ".cache/website",
  target_dir_website = "out/website")
}
\arguments{
\item{target_dir_yaml}{some text}

\item{target_dir_website}{some text}
}
\description{
aswgarsg swegqq
}
\author{
Frederik Sachser
}

